chore(cms): move utilities folder to src for better project organization
