+++ src/app/blocks/Banner/Component.tsx
@@ -1,6 +1,6 @@
 import type { BannerBlock as BannerBlockProps } from "src/payload-types";
 
-import { cn } from "@/app/utilities/cn";
+import { cn } from "@/utilities/cn";
 import React from "react";
 import RichText from "@/components/RichText";
