+++ src/app/components/UI/Label/index.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { cn } from "@/app/utilities/cn";
+import { cn } from "@/utilities/cn";
 import * as LabelPrimitive from "@radix-ui/react-label";
 import { type VariantProps, cva } from "class-variance-authority";
 import * as React from "react";
