+++ src/utilities/generatePreviewPath.ts
